feat: InfoBar rendering and styling enhancements
This commit include implement user registration update functionality in the Firebase. The updated info is now rendered in the infoBar component. To achieve this, two service functions were created in infoBar.services.js - one for updating the information and another for retrieving it. Additionally, made slight styling enhancements to the landing page, including the addition of CSS animation. Also, modufued the style of the forgotPasswordModal and TermsModal.